requirements.txt
python-dotenv
django
django-rest-framework
django-cors-headers
djoser
pillow
stripe